Paths addopted, steps var added, display_size redefined
The size of the screen is made flexible by introducing a pair of
integers in var 'steps' which defines how many steps in x and y
direction there will be.

The arguments of 'gameplay' are modified to include 'steps', instead of
display_size[1], which is now does not play any role. (see simple.py
changes in gameplay class)

The paths that you and I have for apple and snake images are commented
out instead of deleted - so that we only comment out the other's path
and comment in our path when we do our work.