Completed game with classes (a hybrid of simple.py and slither.py)
The new version is made using classes as in simple, but its structure is
modified, and some logic (boundaries and movement) changed.

The important difference is the update() funciton in the class gameplay.
It doesn't need dt variable, if the speed is controlled by FPS of the
clock itself. Variables such as 'cooldown' are expeled compared to
simple, since they seem to be unnecessary.

The growth of snake due to apple eating is realized in an alternative
way compared to both simple and slither, without the use of 'length'
variable. The gowth is implemented literally: the snake's tail grows
when it eats the apple.