Function gameplay.draw() modified to include drawing of an actual image of the snake head
The image of the snake's head is loaded as an argument of
gameplay.draw(). As the head is not invariant under rotations, thus
pointing to a prefered direction, gameplay.draw() function contains a
rotation procedure which rotates the head image depending on the
direction. Then the images is blited to the screen at the position of
the head.

As I don't know how to modify Snake.draw() function to include this
procedure instead, since I do not understand it, the snake head is drawn
on top of the 'head' from the body list, which doesn't look nice. (see
slither_class.py for a work-around of this problem)