Modified to include the image of the snake's head; the update of gameplay.velocity in update() simplified
As the image of the head should be drawn separately from the body, the
function draw_snake()  (former draw_body() ) in the snake class is
modified so that iterative procedure which draws the body draws only the
body parts excluding the head (the last element in snake.body ). Also,
the pygame blit function draws the image of the head on the screen at
current head coordinates; the head image is loaded as an argument of
this function.

As the image of the head has a prefered direction, it has to be rotated
so that the head will always point to the direction of movement. This is
taken care of by pygame.transform.rotate() function in the
gameplay.draw() function. The additional argument of the gameplay.draw()
function loads the image, the rotation procedure happens, and then the
head oriented head image is loaded to the function draw_snake().

Due to two independent drawing functions in draw_snake() function (for
the body and for the head), the order of drawing matters. An issue is
noticed if deletion of the body tails happens before the body is grown
upon the appel-eating: the body part is drawn on top of the head, before
the body part is removed, so the head disappears when it eats the first
apple. Therefore, deletion of body tail (lines 134) is moved after the
apple-eating check and body growth.