Color choice enabled; gameover variable added
The color of the snake and the apple can now be chooses from the
main.py. (see the commit for main.py)

Additionally, the game is over when the snake bites itself. In simple.py
this was not implemented, but argument in gameplay is added
nevertheless. In slither_class.py the argument gameplay.gameover
controls 'the switch' which main.py should take care of by stopping the
game and asking informing the player (not yet implemented). The switch
is True if the snake bites itself.