Choice of apple color substitutted with apple image
The apple is represented by an imported image now, instead of a block
with a chosen color. For this, apple_color argument of gameplay.draw()
is substittuted for an apple image argument.

simple.py will not have apple image, but the color will be coded as red
within simple.py when drawing an apple. (see simple.py)

slither_class.py will now blit the apple image instead of drawing a red
block (see slither_class.py)