An idea for gameover leading to the main screen
The idea is to introduce a point in the gui loop at which the version is
reset. This can be done simply by migrating the declaration of 'version'
variable  <into>  the gui loop, as the first statement.

In this way, when the game over switches switch to False, the program
comes back to reset the version at the begining of the gui loop and the
choice fo version can be done again.

Then no changes in tools are necessary.