Score menu added at the bottom of the screen
Pygame puts origin at upper left corner, which forces the menu to be at
the bottom edge if one wants boundary conditions to remain fixed.

The other solution can be to make boundary conditions depending on all
four edges which are then defined in main.py from the screen size and
called in gameplay.update() as an argument.